Add test and translation.
